requirements.txt
httpx>=0.27,<1
pytr>=0.4.9,<1
cryptography>=43,<45
PyJWT>=2.9,<3
python-dotenv>=1.0,<2
